+++ include/Ark/VM/VM.hpp
@@ -24,7 +24,6 @@
 
 #include <Ark/VM/Value.hpp>
 #include <Ark/VM/Scope.hpp>
-#include <Ark/VM/Frame.hpp>
 #include <Ark/VM/State.hpp>
 #include <Ark/Builtins/Builtins.hpp>
 #include <Ark/Log.hpp>
@@ -78,7 +77,9 @@ namespace Ark
         template <typename... Args>
         internal::Value call(const std::string& name, Args&&... args);
 
-        // function calling from plugins
+        // ================================================
+        //         function calling from plugins
+        // ================================================
 
         /**
          * @brief Resolving a function call (called by plugins)
@@ -118,26 +119,27 @@ namespace Ark
     private:
         State* m_state;
 
-        int m_exitCode;
-        int m_ip;           // instruction pointer
-        std::size_t m_pp;   // page pointer
+        int m_exitCode;     ///< VM exit code, defaults to 0. Can be changed through `sys:exit`
+        int m_ip;           ///< instruction pointer
+        std::size_t m_pp;   ///< page pointer
+        uint16_t m_sp;      ///< stack pointer
+        uint16_t m_fc;      ///< current frames count
         bool m_running;
         uint16_t m_last_sym_loaded;
         std::size_t m_until_frame_count;
         std::mutex m_mutex;
 
         // related to the execution
-        std::vector<internal::Frame> m_frames;
+        std::array<Value, 8192> m_stack;
         std::vector<uint8_t> m_scope_count_to_delete;
         std::optional<internal::Scope_t> m_saved_scope;
         std::vector<internal::Scope_t> m_locals;
         std::vector<std::shared_ptr<internal::SharedLibrary>> m_shared_lib_objects;
 
-        // just a nice little trick for operator[]
+        // just a nice little trick for operator[] and for pop
         internal::Value m__no_value = internal::Builtins::nil;
 
-        // needed to pass data around when binding ArkScript in a program
-        void* m_user_pointer;
+        void* m_user_pointer; ///< needed to pass data around when binding ArkScript in a program
 
         /**
          * @brief Run ArkScript bytecode inside a try catch to retrieve all the exceptions and display a stack trace if needed
@@ -153,7 +155,30 @@ namespace Ark
          */
         void init() noexcept;
 
-        // locals related
+        // ================================================
+        //                 stack related
+        // ================================================
+
+        /**
+         * @brief Pop a value from the stack
+         * 
+         * @return internal::Value* 
+         */
+        inline internal::Value* pop();
+
+        /**
+         * @brief Push a value on the stack
+         * 
+         * @param val 
+         */
+        inline void push(const internal::Value& val);
+
+        /**
+         * @brief Push a value on the stack
+         * 
+         * @param val 
+         */
+        inline void push(internal::Value&& val);
 
         /**
          * @brief Pop a value from the stack and resolve it if possible, then return it
@@ -162,6 +187,10 @@ namespace Ark
          * @return internal::Value* 
          */
         inline internal::Value* popAndResolveAsPtr(int frame=-1);
+
+        // ================================================
+        //                locals related
+        // ================================================
 
         /**
          * @brief Find the nearest variable of a given id
@@ -187,7 +216,9 @@ namespace Ark
          */
         void loadPlugin(uint16_t id);
 
-        // error handling
+        // ================================================
+        //                  error handling
+        // ================================================
 
         /**
          * @brief Find the nearest variable id with a given value